Refactored from MXNet extension to functional interface
