+++ test/sagemaker_tests/mxnet/__init__.py
@@ -10,41 +10,7 @@
 # distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF
 # ANY KIND, either express or implied. See the License for the specific
 # language governing permissions and limitations under the License.
-import os
-import sagemaker
-import boto3
-import subprocess
-import re
-from sagemaker import Session
-from sagemaker.mxnet.estimator import MXNet as MXNet_origin
-from ..pytorch import get_ecr_image_region, get_sagemaker_session, get_ecr_image
-from base64 import b64decode
+from ..sm_multi_region_executor import invoke_sagemaker_test_function
 
-
-class MXNet(MXNet_origin):
-    def __init__(self, image_uri, sagemaker_regions, **kwargs):
-        super().__init__(image_uri, **kwargs)
-        self.image_uri = image_uri
-        self.sagemaker_regions = sagemaker_regions
-
-    def fit(self, inputs=None, wait=True, logs="All", job_name=None, experiment_config=None, **kwargs):
-
-        for region in self.sagemaker_regions:
-            sagemaker_session = get_sagemaker_session(region)
-            # Upload the image to test region if needed
-            if region != get_ecr_image_region(self.image_uri):
-                self.image_uri = get_ecr_image(self.image_uri, region)
-            try:
-                super().fit(
-                    inputs=inputs,
-                    wait=wait,
-                    logs=logs,
-                    job_name=job_name,
-                    experiment_config=experiment_config
-                )
-                return
-            except sagemaker.exceptions.UnexpectedStatusException as e:
-                if "CapacityError" in str(e):
-                    continue
-                else:
-                    raise e
+def invoke_mxnet_helper_function(ecr_image, sagemaker_regions, helper_function, helper_function_args):
+    invoke_sagemaker_test_function(ecr_image, sagemaker_regions, helper_function, helper_function_args)
